pyproject.toml
[tool.poetry]
name = "hojo"
version = "0.1.0"
description = "Hojo: An ORM built on top of SQL Alchemy"
authors = ["Rafael Izidoro <izidoro.me@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = "^2.0.23"
alembic = "^1.13.0"
uuid6 = "^2023.5.2"
pluralizer = "^1.2.0"
pg8000 = "^1.30.3"
stringcase = "^1.2.0"
attrs = "^23.1.0"
cattrs = "^23.2.3"

[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"
black = "^23.10.0"
isort = "^5.12.0"
mypy = "^1.6.1"
pytest = "^7.4.2"
import-deps = "^0.2.0"
safety = "^2.3.5"
bandit = "^1.7.5"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
